Rename src/engine to src/store-engine to make it clearer what it is
